flask_dash_app/requirements.txt
flask
PyJWT
matplotlib
geoip2fast
werkzeug
elasticsearch
elasticsearch_dsl[async]
discord
python-dotenv
Flask-Login 
Flask-SQLAlchemy
Flask-WTF
pandas
scikit-learn
imblearn
dash
setuptools
dash_bootstrap_components
dash_ag_grid
geoip2
plotly==5.24.1
gunicorn
